+++ packages/dialect-react-ui/utils/getExplorerAddress.ts
@@ -1,13 +1,15 @@
+const RELIEABLE_DEVNET = 'devnet-solana';
+
 export const getExplorerAddress = (
   address: string,
   cluster: string | null | undefined
 ): string => {
-  if (!cluster) {
-    cluster = 'devnet';
+  if (!cluster || cluster === 'devnet') {
+    cluster = RELIEABLE_DEVNET;
   } else if (cluster === 'mainnet') {
     cluster = '';
   }
-  return `https://solscan.io/account/${address}${
+  return `https://solana.fm/address/${address}${
     cluster ? `?cluster=${cluster}` : ''
   }`;
 };